organize cookbook into folders
