+++ parea/cookbook/assets/data/anthropic_tool_use_examples.py
