+++ parea/cookbook/assets/data/openai_input_examples.py
