+++ parea/cookbook/openai/simple_experiment_with_openai.py
