+++ parea/cookbook/openai/tracing_open_ai_streams.py
@@ -5,7 +5,7 @@ from dotenv import load_dotenv
 from openai import AsyncOpenAI, OpenAI
 
 from parea import Parea, trace
-from parea.cookbook.data.openai_input_examples import functions_example, simple_example_json
+from parea.cookbook.assets.data.openai_input_examples import functions_example, simple_example_json
 
 load_dotenv()
