+++ parea/cookbook/openai/tracing_openai_assistant_endpoint.py
