+++ parea/cookbook/openai/tracing_templated_llm_calls.py
