+++ parea/cookbook/openai/tracing_with_images_open_ai.py
