+++ parea/cookbook/openai/tracing_with_open_ai_endpoint_directly.py
