+++ parea/cookbook/openai/tracing_with_openai_with_functions.py
